🚚 move: address 관련 파일들 member 패키지 안으로 이동
